% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\docType{data}
\name{.rye_gensym_env}
\alias{.rye_gensym_env}
\title{Global gensym counter (stored in environment to avoid locking issues)}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.rye_gensym_env
}
\description{
Global gensym counter (stored in environment to avoid locking issues)
}
\keyword{datasets}
