% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\docType{data}
\name{.rye_hygiene_gensym_env}
\alias{.rye_hygiene_gensym_env}
\title{Global hygiene gensym counter}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.rye_hygiene_gensym_env
}
\description{
Global hygiene gensym counter
}
\keyword{internal}
