% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\docType{data}
\name{.rye_macros}
\alias{.rye_macros}
\title{Global macro registry}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.rye_macros
}
\description{
Global macro registry
}
\keyword{internal}
