% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{Evaluator}
\alias{Evaluator}
\title{Evaluator for Rye expressions}
\description{
Evaluator for Rye expressions

Evaluator for Rye expressions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Evaluator-new}{\code{Evaluator$new()}}
\item \href{#method-Evaluator-eval}{\code{Evaluator$eval()}}
\item \href{#method-Evaluator-eval_seq}{\code{Evaluator$eval_seq()}}
\item \href{#method-Evaluator-eval_inner}{\code{Evaluator$eval_inner()}}
\item \href{#method-Evaluator-eval_inner_impl}{\code{Evaluator$eval_inner_impl()}}
\item \href{#method-Evaluator-eval_args}{\code{Evaluator$eval_args()}}
\item \href{#method-Evaluator-apply}{\code{Evaluator$apply()}}
\item \href{#method-Evaluator-quote_arg}{\code{Evaluator$quote_arg()}}
\item \href{#method-Evaluator-do_call}{\code{Evaluator$do_call()}}
\item \href{#method-Evaluator-promise_new}{\code{Evaluator$promise_new()}}
\item \href{#method-Evaluator-apply_closure}{\code{Evaluator$apply_closure()}}
\item \href{#method-Evaluator-eval_package_access}{\code{Evaluator$eval_package_access()}}
\item \href{#method-Evaluator-build_special_forms}{\code{Evaluator$build_special_forms()}}
\item \href{#method-Evaluator-clone}{\code{Evaluator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$new(env, macro_expander, source_tracker, engine = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$eval(expr, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-eval_seq"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-eval_seq}{}}}
\subsection{Method \code{eval_seq()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$eval_seq(exprs, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-eval_inner"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-eval_inner}{}}}
\subsection{Method \code{eval_inner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$eval_inner(expr, env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-eval_inner_impl"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-eval_inner_impl}{}}}
\subsection{Method \code{eval_inner_impl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$eval_inner_impl(expr, env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-eval_args"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-eval_args}{}}}
\subsection{Method \code{eval_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$eval_args(expr, env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-apply"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$apply(fn, args, env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-quote_arg"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-quote_arg}{}}}
\subsection{Method \code{quote_arg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$quote_arg(value, quote_symbols = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-do_call"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-do_call}{}}}
\subsection{Method \code{do_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$do_call(fn, args)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-promise_new"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-promise_new}{}}}
\subsection{Method \code{promise_new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$promise_new(expr, env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-apply_closure"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-apply_closure}{}}}
\subsection{Method \code{apply_closure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$apply_closure(fn, args, env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-eval_package_access"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-eval_package_access}{}}}
\subsection{Method \code{eval_package_access()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$eval_package_access(expr, env, op_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-build_special_forms"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-build_special_forms}{}}}
\subsection{Method \code{build_special_forms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$build_special_forms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
