% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{gensym}
\alias{gensym}
\title{Generate a unique symbol}
\usage{
gensym(prefix = "G")
}
\arguments{
\item{prefix}{Optional prefix for the symbol (default: "G")}
}
\value{
A unique symbol
}
\description{
Generate a unique symbol
}
