% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{get_macro}
\alias{get_macro}
\title{Get a macro expander function}
\usage{
get_macro(name, env = parent.frame())
}
\arguments{
\item{name}{Symbol naming the macro}
}
\value{
The macro expander function
}
\description{
Get a macro expander function
}
\keyword{internal}
