% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{is_macro}
\alias{is_macro}
\title{Check if a symbol names a macro}
\usage{
is_macro(name)
}
\arguments{
\item{name}{Symbol to check}
}
\value{
TRUE if it's a macro, FALSE otherwise
}
\description{
Check if a symbol names a macro
}
