% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{MacroExpander}
\alias{MacroExpander}
\title{Macro expander}
\description{
Macro expander

Macro expander
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MacroExpander-new}{\code{MacroExpander$new()}}
\item \href{#method-MacroExpander-defmacro}{\code{MacroExpander$defmacro()}}
\item \href{#method-MacroExpander-macroexpand}{\code{MacroExpander$macroexpand()}}
\item \href{#method-MacroExpander-macroexpand_1}{\code{MacroExpander$macroexpand_1()}}
\item \href{#method-MacroExpander-is_macro}{\code{MacroExpander$is_macro()}}
\item \href{#method-MacroExpander-get_macro}{\code{MacroExpander$get_macro()}}
\item \href{#method-MacroExpander-capture}{\code{MacroExpander$capture()}}
\item \href{#method-MacroExpander-hygiene_unwrap}{\code{MacroExpander$hygiene_unwrap()}}
\item \href{#method-MacroExpander-quasiquote}{\code{MacroExpander$quasiquote()}}
\item \href{#method-MacroExpander-clone}{\code{MacroExpander$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-new"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$new(env, source_tracker = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-defmacro"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-defmacro}{}}}
\subsection{Method \code{defmacro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$defmacro(name, params, body, docstring = NULL, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-macroexpand"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-macroexpand}{}}}
\subsection{Method \code{macroexpand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$macroexpand(expr, env = NULL, preserve_src = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-macroexpand_1"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-macroexpand_1}{}}}
\subsection{Method \code{macroexpand_1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$macroexpand_1(expr, env = NULL, preserve_src = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-is_macro"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-is_macro}{}}}
\subsection{Method \code{is_macro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$is_macro(name, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-get_macro"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-get_macro}{}}}
\subsection{Method \code{get_macro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$get_macro(name, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-capture"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-capture}{}}}
\subsection{Method \code{capture()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$capture(symbol, expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-hygiene_unwrap"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-hygiene_unwrap}{}}}
\subsection{Method \code{hygiene_unwrap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$hygiene_unwrap(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-quasiquote"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-quasiquote}{}}}
\subsection{Method \code{quasiquote()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$quasiquote(expr, env, depth = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MacroExpander-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MacroExpander-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MacroExpander$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
