% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-registry.R
\name{ModuleRegistry}
\alias{ModuleRegistry}
\title{Module registry wrapper}
\description{
Module registry wrapper

Module registry wrapper
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModuleRegistry-new}{\code{ModuleRegistry$new()}}
\item \href{#method-ModuleRegistry-exists}{\code{ModuleRegistry$exists()}}
\item \href{#method-ModuleRegistry-get}{\code{ModuleRegistry$get()}}
\item \href{#method-ModuleRegistry-register}{\code{ModuleRegistry$register()}}
\item \href{#method-ModuleRegistry-update_exports}{\code{ModuleRegistry$update_exports()}}
\item \href{#method-ModuleRegistry-unregister}{\code{ModuleRegistry$unregister()}}
\item \href{#method-ModuleRegistry-attach}{\code{ModuleRegistry$attach()}}
\item \href{#method-ModuleRegistry-clone}{\code{ModuleRegistry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleRegistry-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleRegistry-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleRegistry$new(rye_env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleRegistry-exists"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleRegistry-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleRegistry$exists(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleRegistry-get"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleRegistry-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleRegistry$get(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleRegistry-register"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleRegistry-register}{}}}
\subsection{Method \code{register()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleRegistry$register(name, env, exports, path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleRegistry-update_exports"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleRegistry-update_exports}{}}}
\subsection{Method \code{update_exports()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleRegistry$update_exports(name, exports)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleRegistry-unregister"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleRegistry-unregister}{}}}
\subsection{Method \code{unregister()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleRegistry$unregister(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleRegistry-attach"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleRegistry-attach}{}}}
\subsection{Method \code{attach()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleRegistry$attach(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleRegistry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleRegistry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleRegistry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
