% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{Parser}
\alias{Parser}
\title{Parser for Rye tokens}
\description{
Parser for Rye tokens

Parser for Rye tokens
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Parser-new}{\code{Parser$new()}}
\item \href{#method-Parser-parse}{\code{Parser$parse()}}
\item \href{#method-Parser-clone}{\code{Parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-new"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$new(source_tracker = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse}{}}}
\subsection{Method \code{parse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse(tokens, source_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
