% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{rye_capture}
\alias{rye_capture}
\title{Mark occurrences of a symbol as introduced}
\usage{
rye_capture(symbol, expr)
}
\arguments{
\item{symbol}{A symbol or single string naming the identifier to capture}

\item{expr}{Expression to scan for occurrences}
}
\description{
Mark occurrences of a symbol as introduced
}
\keyword{internal}
