% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{rye_defmacro}
\alias{rye_defmacro}
\title{Define a macro}
\usage{
rye_defmacro(name, params, body, env)
}
\arguments{
\item{name}{Symbol naming the macro}

\item{params}{Parameter list}

\item{body}{Macro body expressions}

\item{env}{Environment in which to define the macro}
}
\description{
Define a macro
}
\keyword{internal}
