% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rye_load_file}
\alias{rye_load_file}
\title{Load and evaluate a Rye source file}
\usage{
rye_load_file(path, env = parent.frame())
}
\arguments{
\item{path}{Path to a Rye source file}

\item{env}{Environment in which to evaluate the file}
}
\value{
The result of the final expression in the file
}
\description{
Load and evaluate a Rye source file
}
