% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rye_load_prelude}
\alias{rye_load_prelude}
\title{Load the Rye prelude into an environment}
\usage{
rye_load_prelude(env = parent.frame())
}
\arguments{
\item{env}{Environment in which to evaluate the prelude}
}
\value{
The result of the final expression in the prelude
}
\description{
Load the Rye prelude into an environment
}
