% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rye_load_stdlib_files}
\alias{rye_load_stdlib_files}
\title{Load all Rye standard library source files into an environment}
\usage{
rye_load_stdlib_files(env = parent.frame())
}
\arguments{
\item{env}{Environment in which to evaluate the standard library files}
}
\value{
The result of the final expression in the last file
}
\description{
Load all Rye standard library source files into an environment
}
