% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{rye_macroexpand}
\alias{rye_macroexpand}
\title{Expand macros in an expression}
\usage{
rye_macroexpand(expr, env = parent.frame())
}
\arguments{
\item{expr}{Expression to expand}

\item{env}{Environment for evaluation}
}
\value{
Expanded expression
}
\description{
Expand macros in an expression
}
