% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{rye_macroexpand_1}
\alias{rye_macroexpand_1}
\title{Expand a single macro layer in an expression}
\usage{
rye_macroexpand_1(expr, env = parent.frame())
}
\arguments{
\item{expr}{Expression to expand}

\item{env}{Environment for evaluation}
}
\value{
Expanded expression or the original expression
}
\description{
Expand a single macro layer in an expression
}
\keyword{internal}
