% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{rye_parse}
\alias{rye_parse}
\title{Parse Rye tokens into R data structures}
\usage{
rye_parse(tokens)
}
\arguments{
\item{tokens}{A list of tokens from rye_tokenize}
}
\value{
A list of parsed expressions as R calls
}
\description{
Parse Rye tokens into R data structures
}
