% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{rye_quasiquote}
\alias{rye_quasiquote}
\title{Process quasiquote expressions}
\usage{
rye_quasiquote(expr, env, depth = 1)
}
\arguments{
\item{expr}{The expression to process}

\item{env}{Environment for evaluation}

\item{depth}{Nesting depth of quasiquote (for nested quasiquotes)}
}
\value{
The processed expression
}
\description{
Process quasiquote expressions
}
