% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{rye_read}
\alias{rye_read}
\title{Read Rye source code into parsed expressions}
\usage{
rye_read(source)
}
\arguments{
\item{source}{A string containing Rye source code}
}
\value{
A list of parsed expressions
}
\description{
Read Rye source code into parsed expressions
}
