% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{rye_tokenize}
\alias{rye_tokenize}
\title{Tokenize Rye source code}
\usage{
rye_tokenize(source)
}
\arguments{
\item{source}{A string containing Rye source code}
}
\value{
A list of tokens, each with type, value, line, and column
}
\description{
Tokenize Rye source code
}
