% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\name{rye_translate}
\alias{rye_translate}
\title{Translate Rye code to R code}
\usage{
rye_translate(source, is_file = NULL)
}
\arguments{
\item{source}{Either a file path to a .rye file or a string containing Rye code}

\item{is_file}{Logical indicating whether source is a file path (default: TRUE if source ends with .rye)}
}
\value{
A character string containing the translated R code
}
\description{
Translate Rye code to R code
}
