% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rye-env.R
\name{RyeEnv}
\alias{RyeEnv}
\title{Rye environment wrapper}
\description{
Rye environment wrapper

Rye environment wrapper
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RyeEnv-new}{\code{RyeEnv$new()}}
\item \href{#method-RyeEnv-raw}{\code{RyeEnv$raw()}}
\item \href{#method-RyeEnv-assign}{\code{RyeEnv$assign()}}
\item \href{#method-RyeEnv-assign_existing}{\code{RyeEnv$assign_existing()}}
\item \href{#method-RyeEnv-resolve_module}{\code{RyeEnv$resolve_module()}}
\item \href{#method-RyeEnv-attach_module}{\code{RyeEnv$attach_module()}}
\item \href{#method-RyeEnv-format_value}{\code{RyeEnv$format_value()}}
\item \href{#method-RyeEnv-symbol_or_string}{\code{RyeEnv$symbol_or_string()}}
\item \href{#method-RyeEnv-macro_registry_env}{\code{RyeEnv$macro_registry_env()}}
\item \href{#method-RyeEnv-module_registry_env}{\code{RyeEnv$module_registry_env()}}
\item \href{#method-RyeEnv-find_define_env}{\code{RyeEnv$find_define_env()}}
\item \href{#method-RyeEnv-find_existing_env}{\code{RyeEnv$find_existing_env()}}
\item \href{#method-RyeEnv-assign_pattern}{\code{RyeEnv$assign_pattern()}}
\item \href{#method-RyeEnv-destructure_bind}{\code{RyeEnv$destructure_bind()}}
\item \href{#method-RyeEnv-clone}{\code{RyeEnv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-new"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$new(env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-raw"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-raw}{}}}
\subsection{Method \code{raw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$raw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-assign"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$assign(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-assign_existing"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-assign_existing}{}}}
\subsection{Method \code{assign_existing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$assign_existing(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-resolve_module"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-resolve_module}{}}}
\subsection{Method \code{resolve_module()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$resolve_module(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-attach_module"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-attach_module}{}}}
\subsection{Method \code{attach_module()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$attach_module(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-format_value"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-format_value}{}}}
\subsection{Method \code{format_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$format_value(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-symbol_or_string"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-symbol_or_string}{}}}
\subsection{Method \code{symbol_or_string()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$symbol_or_string(value, message)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-macro_registry_env"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-macro_registry_env}{}}}
\subsection{Method \code{macro_registry_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$macro_registry_env(env = NULL, create = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-module_registry_env"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-module_registry_env}{}}}
\subsection{Method \code{module_registry_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$module_registry_env(env = NULL, create = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-find_define_env"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-find_define_env}{}}}
\subsection{Method \code{find_define_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$find_define_env(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-find_existing_env"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-find_existing_env}{}}}
\subsection{Method \code{find_existing_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$find_existing_env(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-assign_pattern"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-assign_pattern}{}}}
\subsection{Method \code{assign_pattern()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$assign_pattern(
  pattern,
  value,
  mode = c("define", "set"),
  context = "define"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-destructure_bind"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-destructure_bind}{}}}
\subsection{Method \code{destructure_bind()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$destructure_bind(pattern, value, mode = c("define", "set"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEnv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEnv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEnv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
