% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-tracker.R
\name{SourceTracker}
\alias{SourceTracker}
\title{Source tracking for errors}
\description{
Source tracking for errors

Source tracking for errors
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SourceTracker-src_new}{\code{SourceTracker$src_new()}}
\item \href{#method-SourceTracker-src_get}{\code{SourceTracker$src_get()}}
\item \href{#method-SourceTracker-src_set}{\code{SourceTracker$src_set()}}
\item \href{#method-SourceTracker-src_inherit}{\code{SourceTracker$src_inherit()}}
\item \href{#method-SourceTracker-strip_src}{\code{SourceTracker$strip_src()}}
\item \href{#method-SourceTracker-get}{\code{SourceTracker$get()}}
\item \href{#method-SourceTracker-reset}{\code{SourceTracker$reset()}}
\item \href{#method-SourceTracker-push}{\code{SourceTracker$push()}}
\item \href{#method-SourceTracker-pop}{\code{SourceTracker$pop()}}
\item \href{#method-SourceTracker-format_src}{\code{SourceTracker$format_src()}}
\item \href{#method-SourceTracker-format_error}{\code{SourceTracker$format_error()}}
\item \href{#method-SourceTracker-print_error}{\code{SourceTracker$print_error()}}
\item \href{#method-SourceTracker-with_error_context}{\code{SourceTracker$with_error_context()}}
\item \href{#method-SourceTracker-clone}{\code{SourceTracker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-src_new"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-src_new}{}}}
\subsection{Method \code{src_new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$src_new(
  file,
  start_line,
  start_col,
  end_line = start_line,
  end_col = start_col
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-src_get"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-src_get}{}}}
\subsection{Method \code{src_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$src_get(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-src_set"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-src_set}{}}}
\subsection{Method \code{src_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$src_set(expr, src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-src_inherit"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-src_inherit}{}}}
\subsection{Method \code{src_inherit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$src_inherit(expr, from)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-strip_src"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-strip_src}{}}}
\subsection{Method \code{strip_src()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$strip_src(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-get"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-reset"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-push"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-push}{}}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$push(src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-pop"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-pop}{}}}
\subsection{Method \code{pop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$pop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-format_src"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-format_src}{}}}
\subsection{Method \code{format_src()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$format_src(src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-format_error"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-format_error}{}}}
\subsection{Method \code{format_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$format_error(e, include_r_stack = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-print_error"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-print_error}{}}}
\subsection{Method \code{print_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$print_error(e, file = stderr())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-with_error_context"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-with_error_context}{}}}
\subsection{Method \code{with_error_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$with_error_context(fn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceTracker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SourceTracker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceTracker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
