% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{Tokenizer}
\alias{Tokenizer}
\title{Tokenizer for Rye source code}
\description{
Tokenizer for Rye source code

Tokenizer for Rye source code
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tokenizer-new}{\code{Tokenizer$new()}}
\item \href{#method-Tokenizer-tokenize}{\code{Tokenizer$tokenize()}}
\item \href{#method-Tokenizer-clone}{\code{Tokenizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tokenizer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tokenizer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tokenizer$new(delimiters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tokenizer-tokenize"></a>}}
\if{latex}{\out{\hypertarget{method-Tokenizer-tokenize}{}}}
\subsection{Method \code{tokenize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tokenizer$tokenize(source)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tokenizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tokenizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tokenizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
