% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{rye_eval}
\alias{rye_eval}
\title{Evaluate Rye expressions}
\usage{
rye_eval(expr, env = parent.frame())
}
\arguments{
\item{expr}{A parsed Rye expression}

\item{env}{Environment in which to evaluate (default: caller's environment)}
}
\value{
The result of evaluation
}
\description{
Evaluate Rye expressions
}
