% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdlib.R
\name{rye_load_stdlib}
\alias{rye_load_stdlib}
\title{Load the Rye standard library}
\usage{
rye_load_stdlib(env = .GlobalEnv)
}
\arguments{
\item{env}{Environment in which to load the library}
}
\description{
Load the Rye standard library
}
