% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{rye_macroexpand}
\alias{rye_macroexpand}
\title{Expand macros in an expression}
\usage{
rye_macroexpand(expr, env = parent.frame())
}
\arguments{
\item{expr}{Expression to expand}

\item{env}{Environment for evaluation}
}
\value{
Expanded expression
}
\description{
Expands macros recursively until the expression no longer begins with a macro
call. Quoted and quasiquoted forms are not expanded.
}
\examples{
rye_eval(rye_read("(defmacro when (test body) `(if ,test ,body #nil))")[[1]])
rye_macroexpand(rye_read("(when #t 1)")[[1]])
}
