% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rye-engine.R
\name{RyeEngine}
\alias{RyeEngine}
\title{Core Rye engine}
\description{
Provides class-based organization for tokenization, parsing, macro expansion,
evaluation, and environment management.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RyeEngine-new}{\code{RyeEngine$new()}}
\item \href{#method-RyeEngine-read}{\code{RyeEngine$read()}}
\item \href{#method-RyeEngine-tokenize}{\code{RyeEngine$tokenize()}}
\item \href{#method-RyeEngine-parse}{\code{RyeEngine$parse()}}
\item \href{#method-RyeEngine-eval}{\code{RyeEngine$eval()}}
\item \href{#method-RyeEngine-eval_seq}{\code{RyeEngine$eval_seq()}}
\item \href{#method-RyeEngine-eval_exprs}{\code{RyeEngine$eval_exprs()}}
\item \href{#method-RyeEngine-eval_text}{\code{RyeEngine$eval_text()}}
\item \href{#method-RyeEngine-initialize_environment}{\code{RyeEngine$initialize_environment()}}
\item \href{#method-RyeEngine-load_file}{\code{RyeEngine$load_file()}}
\item \href{#method-RyeEngine-load_stdlib}{\code{RyeEngine$load_stdlib()}}
\item \href{#method-RyeEngine-macroexpand}{\code{RyeEngine$macroexpand()}}
\item \href{#method-RyeEngine-macroexpand_1}{\code{RyeEngine$macroexpand_1()}}
\item \href{#method-RyeEngine-repl}{\code{RyeEngine$repl()}}
\item \href{#method-RyeEngine-translate}{\code{RyeEngine$translate()}}
\item \href{#method-RyeEngine-clone}{\code{RyeEngine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-new"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$new(env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-read"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$read(source, source_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-tokenize"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-tokenize}{}}}
\subsection{Method \code{tokenize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$tokenize(source)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-parse"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-parse}{}}}
\subsection{Method \code{parse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$parse(tokens, source_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-eval"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$eval(expr, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-eval_seq"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-eval_seq}{}}}
\subsection{Method \code{eval_seq()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$eval_seq(exprs, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-eval_exprs"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-eval_exprs}{}}}
\subsection{Method \code{eval_exprs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$eval_exprs(exprs, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-eval_text"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-eval_text}{}}}
\subsection{Method \code{eval_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$eval_text(text, env = NULL, source_name = "<eval>")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-initialize_environment"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-initialize_environment}{}}}
\subsection{Method \code{initialize_environment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$initialize_environment(env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-load_file"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-load_file}{}}}
\subsection{Method \code{load_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$load_file(path, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-load_stdlib"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-load_stdlib}{}}}
\subsection{Method \code{load_stdlib()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$load_stdlib(env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-macroexpand"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-macroexpand}{}}}
\subsection{Method \code{macroexpand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$macroexpand(expr, env = NULL, preserve_src = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-macroexpand_1"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-macroexpand_1}{}}}
\subsection{Method \code{macroexpand_1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$macroexpand_1(expr, env = NULL, preserve_src = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-repl"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-repl}{}}}
\subsection{Method \code{repl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$repl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-translate"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-translate}{}}}
\subsection{Method \code{translate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$translate(source, is_file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RyeEngine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RyeEngine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RyeEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
